+++ afat/views/__init__.py
@@ -1,5 +1,3 @@
 """
 initialize the views
 """
-
-from afat.views import dashboard, statistics, fatlinks  # noqa: F401